backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pymongo
numpy
slowapi
pytest
redis
rq
PyJWT
